Fix font identification error when fonts are not installed.
If a single font matches the font stack, then check whether the rendered
font is really installed. This is necessary because if none of the fonts in the
stack are installed then the rendered font will always be the browser's default
font, so it will match the font stack even though the font is not installed.